requirements.txt
flask==3.0.3
python-dotenv==1.0.1
langchain==0.3.11
langchain-google-genai==2.0.8
langchain-core==0.3.29
twilio==9.4.0
requests==2.32.3
pydantic==2.10.5
python-dateutil==2.9.0 